pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "research_agent"
version = "0.1.0"
description = "Local-first researcher agent with optional cloud librarian hop"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "pydantic",
  "pyyaml",
  "pytest",
  "rich",
  "faiss-cpu",
  "sentence-transformers",
  "requests",
  "tqdm",
]

[tool.setuptools]
packages = ["researcher"]
